Narrow `.box`es to give Firefox enough space to do multi-column layout.
Sadly, Gecko and WebKit calculate widths slightly differently, and this
is a known problem with the masonry plugin for jQuery that we're using.
In order to solve the problem quickly, we need to give the calculation
with the wider result (Firefox/Gecko) enough space to do its
multi-column layout, else we'll get a single column unintentinoally. For
more information, see:

    https://github.com/desandro/masonry/issues/166